requirements.txt
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.25.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
nltk>=3.8.0
spacy>=3.6.0
xgboost>=1.7.0
lightgbm>=4.0.0
optuna>=3.2.0
shap>=0.42.0
lime>=0.2.0
textstat>=0.7.0
pygments>=2.15.0
joblib>=1.3.0
